requirements/local.txt
-r base.txt
black
factory-boy
pytest
pytest-django
